Move spec support to spec/support
