+++ cobra_commander/spec/support/cobra_commander-stub.gemspec
