+++ cobra_commander/spec/support/stub_source.rb
